app/src/main/java/com/example/yudinata/footballsc/feature/matchs/nextmatch/NextsPresenter.kt
package com.example.yudinata.footballsc.feature.matchs.nextmatch


import com.example.yudinata.footballsc.model.EventResponse
import com.example.yudinata.footballsc.network.ApiRepository
import com.example.yudinata.footballsc.network.TheSportDBApi
import com.example.yudinata.footballsc.util.CoroutineContextProvider
import com.google.gson.Gson
import kotlinx.coroutines.experimental.async
import org.jetbrains.anko.coroutines.experimental.bg

class NextsPresenter(private val view: NextsView,
                     private val apiRepository: ApiRepository,
                     private val gson: Gson, private val context: CoroutineContextProvider = CoroutineContextProvider()) {

    fun getTeamList(league: String?) {
        view.showLoading()
        async(context.main){
            val data = bg{
                gson.fromJson(apiRepository
                        .doRequest(TheSportDBApi.getTeams(league)),
                        EventResponse::class.java)
            }
            view.showTeamList(data.await().events)
            view.hideLoading()

        }
    }
}